python/Pipeline/requirements.txt
absl-py==1.1.0
attrs==19.3.0
Automat==0.8.0
bcrypt==3.2.2
blinker==1.4
cachetools==4.2.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
cloud-init==21.2
cloudpickle==2.1.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
Deprecated==1.2.13
distro==1.4.0
distro-info===0.23ubuntu1
docstring-parser==0.14.1
entrypoints==0.3
fire==0.4.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-storage==1.44.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
httplib2==0.20.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kfp==1.8.12
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.2
kubernetes==18.20.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
pandas==1.4.2
paramiko==2.11.0
pexpect==4.6.0
protobuf==3.20.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==1.9.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.8
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
ssh-import-id==5.10
strip-hints==0.1.10
systemd-python==234
tabulate==0.8.10
termcolor==1.1.0
Twisted==18.9.0
typer==0.5.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
wadllib==1.3.3
websocket-client==1.3.3
wrapt==1.14.1
zipp==1.0.0
zope.interface==4.7.1
